Moved UserInfo to types.ts and updated references
