+++ components/panels/SectionCheckBox.tsx
@@ -12,7 +12,7 @@ import Keys from '../Keys';
 import macros from '../macros';
 import { Section } from '../types';
 import axios from 'axios';
-import { UserInfo } from '../Header';
+import { UserInfo } from '../../components/types';
 
 type SectionCheckBoxProps = {
   section: Section;
